+++ front-end/src/pages/user/common/Header/components/HeaderSignUserRight.jsx
@@ -2,15 +2,15 @@ import React from "react";
 import IconBtn from "components/common/Button/icon/IconBtn";
 import { CiLocationOn, CiSearch, CiShoppingCart } from "react-icons/ci";
 import { Link } from "react-router-dom";
-import HeaderToggle from "components/common/Header/components/HeaderToggle";
-import useToggleDisplay from "components/common/Header/hooks/useToggleDisplay";
+import HeaderToggle from "pages/user/common/Header/components/HeaderToggle";
+import useToggleDisplay from "pages/user/common/Header/hooks/useToggleDisplay";
 import { useRecoilValue } from "recoil";
-import { userInfoState } from "../../../../recoil/user/atoms";
+import { userInfoState } from "../../../../../recoil/user/atoms";
 import {
   HeaderRight,
   LocationBtnWrap,
-} from "components/common/Header/components/HeaderSignUserRightLayout";
-import { Font } from "../../../../assets/styles/CommonStyle";
+} from "pages/user/common/Header/components/HeaderSignUserRightLayout";
+import { Font } from "../../../../../assets/styles/CommonStyle";
 
 function HeaderSignUserRight({ setIsModalOpen, setIsSearchOpen, searchBtnRef }) {
   const [menuRef, showMenu, hideMenu] = useToggleDisplay();
