+++ front-end/src/pages/user/common/Header/hooks/useSearchBoxSlide.js
